+++ corelib/console.c
@@ -494,7 +494,7 @@ is_data:
 
 		if (pos == sizeof(buf)) {
 			pthread_setcancelstate(PTHREAD_CANCEL_DISABLE, NULL);
-			cw_log(CW_LOG_ERROR, "Console got an overlong line (> %lu bytes!)\n", sizeof(buf));
+			cw_log(CW_LOG_ERROR, "Console got an overlong line (> %lu bytes!)\n", (unsigned long)sizeof(buf));
 			break;
 		}
 	} while (nresp);
